requirements-backend.txt
# ===== LogPilot 后端完整依赖 (Phase 2 + 3) =====

# ---- Phase 1: 核心 ----
streamlit>=1.30
pandas>=2.0
openai>=1.0
httpx>=0.24
python-docx>=1.1.0
pypdf>=4.0

# ---- Phase 2: FastAPI 后端 ----
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9

# ---- Phase 2: Celery 任务队列 ----
celery[redis]>=5.3
redis>=5.0

# ---- Phase 3: RAG 向量检索 (可选) ----
# chromadb>=0.4
# sentence-transformers>=2.2

# ---- Phase 3: 报告导出 ----
# (csv/json/html 使用标准库, 无额外依赖)

